Fixed "merging" of default + function attributes of NumberHelper.
This commit fixes a bug where attributes and textAttributes that were
meant to be set on the underlying \NumberFormatter class weren't
correctly set. It changes the merging of the attributes from array_merge
to '+', the standard union of arrays. This is important because
array_merge renumbers numeric array keys and the class constants that
represent the settings of \NumberFormatter are integers [1].

[1] http://php.net/array_merge