Accept 'fraction_digits' as argument option for \NumberFormatter::FRACTION_DIGITS
The Twig templating engine does not allow the use of variables as keys of
arrays. This implies that number_format* filters could not receive
NumberFormatter options in a nice matter.

This commit enables the use of string representations of the name of the class
constants of NumberFormatter in the attributes arrays.