fix(extension): do not resolve the kernel.default_locale parameter
Resolving this parameter in the extension implies it has to be defined before.
Therefore, it force a certain order on the loaded bundles (basically FrameworkBundle must be loaded before this bundle, since it is the one that defines this parameter).
We should not resolve parameter here but instead let them be resolved the latest possible.